+++ R/on-load.R
@@ -12,7 +12,7 @@
 }
 
 .onUnload <- function(libpath) {
-  if (os_type() != "windows") .Call(c_processx__killem_all)
+  .Call(c_processx__unload_cleanup)
   supervisor_reset()
 }