+++ src/win/processx-win.h
@@ -41,5 +41,6 @@ void processx__error(const char *message, DWORD errorcode, const char *file, int
 void processx__stdio_noinherit(BYTE* buffer);
 int processx__stdio_verify(BYTE* buffer, WORD size);
 double processx__create_time(HANDLE process);
+extern HANDLE processx__connection_iocp;
 
 #endif