+++ tests/testthat/helper.R
@@ -55,3 +55,7 @@ get_pid_by_name_windows <- function(name) {
 get_pid_by_name_linux <- function(name) {
   ## TODO
 }
+
+skip_in_covr <- function() {
+  if (Sys.getenv("R_COVR", "") == "true") skip("in covr")
+}