+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # dev
 
+* Fix a crash on Windows, when a connection that has a pending read internally
+  is finalized.
+
 # processx 3.3.0
 
 * `process` can now redirect the standard error to the standard output, via